refactor: 클래스 이름 수정 및 패키지 분리
