+++ src/main/java/racingcar/view/RacingGameView.java
@@ -1,7 +1,8 @@
-package racingcar;
+package racingcar.view;
 
 import java.util.Scanner;
 import java.util.List;
+import racingcar.domain.Car;
 
 public class RacingGameView {
